pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biblio-review"
version = "0.1.0"
description = "Reproducible bibliometric review pipeline with AI-assisted agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Fernando", email = "fernando@example.com"},
]
keywords = ["bibliometrics", "systematic-review", "science-mapping", "AI"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pandas>=2.1",
    "rispy>=0.9",
    "bibtexparser>=1.4,<2.0",
    "networkx>=3.2",
    "rapidfuzz>=3.5",
    "anthropic>=0.40",
    "rich>=13.7",
    "matplotlib>=3.8",
    "scipy>=1.12",
    "pyalex>=0.14",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.9",
]
viz = [
    "plotly>=5.18",
    "pyvis>=0.3",
    "wordcloud>=1.9",
]

[project.scripts]
biblio-review = "biblio_review.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
